+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Innmind\Doctrine\Sort`
+- `Innmind\Doctrine\Matching`
+
 ### Changed
 
+- `Innmind\Doctrine\Repository::matching()` now returns an instance of `Innmind\Doctrine\Matching`
+- `Innmind\Doctrine\Repository::all()` now returns an instance of `Innmind\Doctrine\Matching`
 - Update `innmind/specification` to version `~3.0`
 
 ### Removed